+++ src/gameobjects/bitmaptext/GetBitmapTextSize.js
@@ -26,6 +26,7 @@ var GetBitmapTextSize = function (src, round)
 
     var chars = src.fontData.chars;
     var lineHeight = src.fontData.lineHeight;
+    var letterSpacing = src.letterSpacing;
 
     var xAdvance = 0;
     var yAdvance = 0;
@@ -98,7 +99,7 @@ var GetBitmapTextSize = function (src, round)
             bh = gh;
         }
 
-        xAdvance += glyph.xAdvance;
+        xAdvance += glyph.xAdvance + letterSpacing;
         indexCount += 1;
         lastGlyph = glyph;
         lastCharCode = charCode;