+++ src/gameobjects/bitmaptext/static/BitmapText.js
@@ -33,6 +33,7 @@ var Render = require('./BitmapTextRender');
  * @property {string} font - [description]
  * @property {string} text - [description]
  * @property {number} fontSize - [description]
+ * @property {number} letterSpacing - Adds/Removes spacing between characters
  */
 
 /**
@@ -128,6 +129,16 @@ var BitmapText = new Class({
          * @since 3.0.0
          */
         this.fontSize = size || this.fontData.size;
+
+        /**
+         * Adds/Removes spacing between characters
+         * Can be a negative or positive number
+         *
+         * @name Phaser.GameObjects.BitmapText#letterSpacing
+         * @type {number}
+         * @since 3.4.0
+         */
+        this.letterSpacing = 0;
 
         this.setTexture(entry.texture, entry.frame);
         this.setPosition(x, y);