+++ src/renderer/webgl/pipelines/TextureTintPipeline.js
@@ -946,6 +946,7 @@ var TextureTintPipeline = new Class({
         var rotation = -bitmapText.rotation;
         var scaleX = bitmapText.scaleX;
         var scaleY = bitmapText.scaleY;
+        var letterSpacing = bitmapText.letterSpacing;
         var sr = Math.sin(rotation);
         var cr = Math.cos(rotation);
         var sra = cr * scaleX;
@@ -1005,7 +1006,7 @@ var TextureTintPipeline = new Class({
                 x += (kerningOffset !== undefined) ? kerningOffset : 0;
             }
 
-            xAdvance += glyph.xAdvance;
+            xAdvance += glyph.xAdvance + letterSpacing;
             indexCount += 1;
             lastGlyph = glyph;
             lastCharCode = charCode;
@@ -1178,6 +1179,7 @@ var TextureTintPipeline = new Class({
         var rotation = -bitmapText.rotation;
         var scaleX = bitmapText.scaleX;
         var scaleY = bitmapText.scaleY;
+        var letterSpacing = bitmapText.letterSpacing;
         var sr = Math.sin(rotation);
         var cr = Math.cos(rotation);
         var sra = cr * scaleX;
@@ -1252,7 +1254,7 @@ var TextureTintPipeline = new Class({
                 x += (kerningOffset !== undefined) ? kerningOffset : 0;
             }
 
-            xAdvance += glyph.xAdvance;
+            xAdvance += glyph.xAdvance + letterSpacing;
             indexCount += 1;
             lastGlyph = glyph;
             lastCharCode = charCode;